pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apple-git"
version = "0.1.0"
description = "Bridge Apple Reminders to GitHub Issues and Pull Requests"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "cypher"}]
keywords = ["macos", "reminders", "github", "automation", "kanban"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: MacOS X",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.26.0",
    "PyGithub>=1.59.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
]

[project.scripts]
apple-git = "apple_git.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
